+++ test/coinkey.test.js
@@ -1,8 +1,7 @@
+var assert = require('assert')
 var CoinKey = require('../')
 var conv = require('binstring')
 var secureRandom = require('secure-random')
-
-require('terst')
 
 describe('CoinKey', function() {
   describe('+ CoinKey()', function() {
@@ -11,15 +10,15 @@ describe('CoinKey', function() {
         var ck1 = new CoinKey()
         var ck2 = new CoinKey()
 
-        EQ (ck1.privateKey.length, 32)
-        EQ (ck2.privateKey.length, 32)
+        assert.equal(ck1.privateKey.length, 32)
+        assert.equal(ck2.privateKey.length, 32)
 
-        NEQ (ck1.privateKey.toString('hex'), ck2.privateKey.toString('hex'))
+        assert.notEqual(ck1.privateKey.toString('hex'), ck2.privateKey.toString('hex'))
 
-        EQ (ck1.compressed, false)
+        assert.equal(ck1.compressed, false)
         CoinKey.compressByDefault = true
         var ck3 = new CoinKey()
-        EQ (ck3.compressed, true)
+        assert.equal(ck3.compressed, true)
         CoinKey.compressByDefault = false //<-- restore default preference
       })
     })
@@ -29,17 +28,17 @@ describe('CoinKey', function() {
         it('should use defaults for rest', function() {
           var privateKey = secureRandom(32)
           var ck1 = new CoinKey(privateKey)
-          EQ (ck1.privateKey.toString('hex'), new Buffer(privateKey).toString('hex'))
+          assert.equal(ck1.privateKey.toString('hex'), new Buffer(privateKey).toString('hex'))
         })
       })
 
       describe('> when versions are passed', function() {
         it('should use the versions', function() {
           var ck = new CoinKey({private: 0x34, public: 0xB4}) //<-- namecoin
-          EQ (ck.privateKey.length, 32)
-          F (ck.compressed)
-          EQ (ck.versions.private, 0x34)
-          EQ (ck.versions.public, 0xB4)
+          assert.equal(ck.privateKey.length, 32)
+          assert(!ck.compressed)
+          assert.equal(ck.versions.private, 0x34)
+          assert.equal(ck.versions.public, 0xB4)
         })
       })
     })
@@ -48,8 +47,8 @@ describe('CoinKey', function() {
       describe('> when compressed and versions are passed', function() {
         it('should generate a random private key', function() {
           var ck = new CoinKey(true, {private: 0x34, public: 0xB4}) //<-- namecoin
-          EQ (ck.privateKey.length, 32)
-          T (ck.compressed)
+          assert.equal(ck.privateKey.length, 32)
+          assert(ck.compressed)
         })
       })
     })
@@ -63,7 +62,7 @@ describe('CoinKey', function() {
         it('should return the uncompressed Bitcoin address', function() {
           var privateKeyHex = "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd"
           var ck = new CoinKey(conv(privateKeyHex, {in: 'hex', out: 'buffer'}), false)
-          EQ (ck.privateWif, '5Hx15HFGyep2CfPxsJKe2fXJsCVn5DEiyoeGGF6JZjGbTRnqfiD')
+          assert.equal(ck.privateWif, '5Hx15HFGyep2CfPxsJKe2fXJsCVn5DEiyoeGGF6JZjGbTRnqfiD')
         })
       })
 
@@ -71,7 +70,7 @@ describe('CoinKey', function() {
         it('should return the compressed Bitcoin address', function() {
           var privateKeyHex = "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd"
           var ck = new CoinKey(conv(privateKeyHex, {in: 'hex', out: 'buffer'}), true)
-          EQ (ck.privateWif, 'KwomKti1X3tYJUUMb1TGSM2mrZk1wb1aHisUNHCQXTZq5auC2qc3')
+          assert.equal(ck.privateWif, 'KwomKti1X3tYJUUMb1TGSM2mrZk1wb1aHisUNHCQXTZq5auC2qc3')
         })
       })
     })
@@ -81,7 +80,7 @@ describe('CoinKey', function() {
         it('should return the uncompressed Bitcoin address', function() {
           var privateKeyHex = "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd"
           var ck = new CoinKey(conv(privateKeyHex, {in: 'hex', out: 'buffer'}), false, {private: 0xB0, public: 0x30})
-          EQ (ck.privateWif, '6uFjYQnot5Gtg3HpP87bp4JUpg4FH1gkkV3RyS7LHBbD9Hpt1na')
+          assert.equal(ck.privateWif, '6uFjYQnot5Gtg3HpP87bp4JUpg4FH1gkkV3RyS7LHBbD9Hpt1na')
         })
       })
 
@@ -89,7 +88,7 @@ describe('CoinKey', function() {
         it('should return the compressed Bitcoin address', function() {
           var privateKeyHex = "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd"
           var ck = new CoinKey(conv(privateKeyHex, {in: 'hex', out: 'buffer'}), true, {private: 0xB0, public: 0x30})
-          EQ (ck.privateWif, 'T3e2me1BvRs95K7E8eQ8eha9oRPL1g2U6vmjE5px6RjzbUTvKZsf')
+          assert.equal(ck.privateWif, 'T3e2me1BvRs95K7E8eQ8eha9oRPL1g2U6vmjE5px6RjzbUTvKZsf')
         })
       })
     })
@@ -101,8 +100,8 @@ describe('CoinKey', function() {
         var privateKeyHex = "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd"
         var hash160Hex = "3c176e659bea0f29a3e9bf7880c112b1b31b4dc8"
         var key = new CoinKey(conv(privateKeyHex, {in: 'hex', out: 'bytes'}), false)
-        EQ (key.publicHash.toString('hex'), hash160Hex)
-        EQ (key.pubKeyHash.toString('hex'), hash160Hex)
+        assert.equal(key.publicHash.toString('hex'), hash160Hex)
+        assert.equal(key.pubKeyHash.toString('hex'), hash160Hex)
       })
     })
 
@@ -111,8 +110,8 @@ describe('CoinKey', function() {
         var privateKeyHex = "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd"
         var hash160Hex = "a1c2f92a9dacbd2991c3897724a93f338e44bdc1"
         var key = new CoinKey(conv(privateKeyHex, {in: 'hex', out: 'bytes'}), true)
-        EQ (key.publicHash.toString('hex'), hash160Hex)
-        EQ (key.pubKeyHash.toString('hex'), hash160Hex)
+        assert.equal(key.publicHash.toString('hex'), hash160Hex)
+        assert.equal(key.pubKeyHash.toString('hex'), hash160Hex)
       })
     })
   })
@@ -123,7 +122,7 @@ describe('CoinKey', function() {
         it('should return the uncompressed Bitcoin address', function() {
           var privateKeyHex = "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd"
           var ck = new CoinKey(conv(privateKeyHex, {in: 'hex', out: 'buffer'}), false)
-          EQ (ck.publicAddress, '16UjcYNBG9GTK4uq2f7yYEbuifqCzoLMGS')
+          assert.equal(ck.publicAddress, '16UjcYNBG9GTK4uq2f7yYEbuifqCzoLMGS')
         })
       })
 
@@ -131,7 +130,7 @@ describe('CoinKey', function() {
         it('should return the compressed Bitcoin address', function() {
           var privateKeyHex = "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd"
           var ck = new CoinKey(conv(privateKeyHex, {in: 'hex', out: 'buffer'}), true)
-          EQ (ck.publicAddress, '1FkKMsKNJqWSDvTvETqcCeHcUQQ64kSC6s')
+          assert.equal(ck.publicAddress, '1FkKMsKNJqWSDvTvETqcCeHcUQQ64kSC6s')
         })
       })
     })
@@ -141,7 +140,7 @@ describe('CoinKey', function() {
         it('should return the uncompressed Bitcoin address', function() {
           var privateKeyHex = "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd"
           var ck = new CoinKey(conv(privateKeyHex, {in: 'hex', out: 'buffer'}), false, {private: 0xB0, public: 0x30})
-          EQ (ck.publicAddress, 'LQhgskg1LoWWZsbzCo7GpFffvtCV8Z5GKZ')
+          assert.equal(ck.publicAddress, 'LQhgskg1LoWWZsbzCo7GpFffvtCV8Z5GKZ')
         })
       })
 
@@ -149,7 +148,7 @@ describe('CoinKey', function() {
         it('should return the compressed Bitcoin address', function() {
           var privateKeyHex = "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd"
           var ck = new CoinKey(conv(privateKeyHex, {in: 'hex', out: 'buffer'}), true, {private: 0xB0, public: 0x30})
-          EQ (ck.publicAddress, 'LZyGd5dCPVkVUjA5QbpuUfMNgcmNDLjswH')
+          assert.equal(ck.publicAddress, 'LZyGd5dCPVkVUjA5QbpuUfMNgcmNDLjswH')
         })
       })
     })
@@ -161,13 +160,13 @@ describe('CoinKey', function() {
         var privateKeyHex = "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd"
         var ck = new CoinKey(new Buffer(privateKeyHex, 'hex'), true)
 
-        EQ (ck.privateWif, 'KwomKti1X3tYJUUMb1TGSM2mrZk1wb1aHisUNHCQXTZq5auC2qc3')
-        EQ (ck.publicAddress, '1FkKMsKNJqWSDvTvETqcCeHcUQQ64kSC6s')
+        assert.equal(ck.privateWif, 'KwomKti1X3tYJUUMb1TGSM2mrZk1wb1aHisUNHCQXTZq5auC2qc3')
+        assert.equal(ck.publicAddress, '1FkKMsKNJqWSDvTvETqcCeHcUQQ64kSC6s')
 
         ck.versions = {public: 0x1E, private: 0x9E} //change to DOGECOIN
 
-        EQ (ck.privateWif, 'QPCgUjWzmfNfXzsQBHJ4KZsPKbmaz99PAyZP9ubFFpBBXWuSQh6n')
-        EQ (ck.publicAddress, 'DKtQu8G1cFQikveWy3qAkQTDMY8PKVU18Z')
+        assert.equal(ck.privateWif, 'QPCgUjWzmfNfXzsQBHJ4KZsPKbmaz99PAyZP9ubFFpBBXWuSQh6n')
+        assert.equal(ck.publicAddress, 'DKtQu8G1cFQikveWy3qAkQTDMY8PKVU18Z')
     
       })
     })
@@ -177,15 +176,15 @@ describe('CoinKey', function() {
         var privateKeyHex = "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd"
         var ck = new CoinKey(new Buffer(privateKeyHex, 'hex'), true)
 
-        EQ (ck.privateWif, 'KwomKti1X3tYJUUMb1TGSM2mrZk1wb1aHisUNHCQXTZq5auC2qc3')
-        EQ (ck.publicAddress, '1FkKMsKNJqWSDvTvETqcCeHcUQQ64kSC6s')
+        assert.equal(ck.privateWif, 'KwomKti1X3tYJUUMb1TGSM2mrZk1wb1aHisUNHCQXTZq5auC2qc3')
+        assert.equal(ck.publicAddress, '1FkKMsKNJqWSDvTvETqcCeHcUQQ64kSC6s')
 
         //change to DOGECOIN
         ck.versions.public = 0x1E
         ck.versions.private = 0x9E 
 
-        EQ (ck.privateWif, 'QPCgUjWzmfNfXzsQBHJ4KZsPKbmaz99PAyZP9ubFFpBBXWuSQh6n')
-        EQ (ck.publicAddress, 'DKtQu8G1cFQikveWy3qAkQTDMY8PKVU18Z')
+        assert.equal(ck.privateWif, 'QPCgUjWzmfNfXzsQBHJ4KZsPKbmaz99PAyZP9ubFFpBBXWuSQh6n')
+        assert.equal(ck.publicAddress, 'DKtQu8G1cFQikveWy3qAkQTDMY8PKVU18Z')
       })
     })
   })
@@ -195,7 +194,7 @@ describe('CoinKey', function() {
     it('should return the hex string of the privateKey', function() {
       var privateKeyHex = "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd"
       var ck = new CoinKey(new Buffer(privateKeyHex, 'hex'))
-      EQ (ck.toString(), ck.privateKey.toString('hex'))
+      assert.equal(ck.toString(), ck.privateKey.toString('hex'))
     })
   })
 
@@ -205,18 +204,18 @@ describe('CoinKey', function() {
       describe('> when compressed', function() {
         it('should create a new compressed CoinKey', function() {
           var ck = CoinKey.fromWif('KwomKti1X3tYJUUMb1TGSM2mrZk1wb1aHisUNHCQXTZq5auC2qc3')
-          EQ (ck.compressed, true)
-          EQ (ck.privateKey.toString('hex'), "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd")
-          EQ (ck.publicAddress, "1FkKMsKNJqWSDvTvETqcCeHcUQQ64kSC6s")
+          assert.equal(ck.compressed, true)
+          assert.equal(ck.privateKey.toString('hex'), "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd")
+          assert.equal(ck.publicAddress, "1FkKMsKNJqWSDvTvETqcCeHcUQQ64kSC6s")
         })
       })
 
       describe('> when not compressed', function() {
         it('should create a new compressed CoinKey', function() {
           var ck = CoinKey.fromWif('5Hx15HFGyep2CfPxsJKe2fXJsCVn5DEiyoeGGF6JZjGbTRnqfiD')
-          EQ (ck.compressed, false)
-          EQ (ck.privateKey.toString('hex'), "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd")
-          EQ (ck.publicAddress, '16UjcYNBG9GTK4uq2f7yYEbuifqCzoLMGS')
+          assert.equal(ck.compressed, false)
+          assert.equal(ck.privateKey.toString('hex'), "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd")
+          assert.equal(ck.publicAddress, '16UjcYNBG9GTK4uq2f7yYEbuifqCzoLMGS')
         })
       })
     })
@@ -225,18 +224,18 @@ describe('CoinKey', function() {
       describe('> when compressed', function() {
         it('should create a new compressed CoinKey', function() {
           var ck = CoinKey.fromWif('TdxZ4HSnd3XgyPAJLwyhqPx6oABwMM7NkhGFTuDL4LkqioDbkZWj')
-          EQ (ck.compressed, true)
-          EQ (ck.privateKey.toString('hex'), "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd")
-          EQ (ck.publicAddress, "NBKgZWpMEDbzkTiRWHABRASXCdo8zWM8qC")
+          assert.equal(ck.compressed, true)
+          assert.equal(ck.privateKey.toString('hex'), "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd")
+          assert.equal(ck.publicAddress, "NBKgZWpMEDbzkTiRWHABRASXCdo8zWM8qC")
         })
       })
 
       describe('> when not compressed', function() {
         it('should create a new compressed CoinKey', function() {
           var ck = CoinKey.fromWif('732iFLawHmyoDKce1SbbdLhaV8X7sq4Fu3Krx2rvFocm7qNwwiQ')
-          EQ (ck.compressed, false)
-          EQ (ck.privateKey.toString('hex'), "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd")
-          EQ (ck.publicAddress, 'N246pBsABXN1qcALJUSYkkkpSuEG3KTTGR')
+          assert.equal(ck.compressed, false)
+          assert.equal(ck.privateKey.toString('hex'), "1184cd2cdd640ca42cfc3a091c51d549b2f016d454b2774019c2b2d2e08529fd")
+          assert.equal(ck.publicAddress, 'N246pBsABXN1qcALJUSYkkkpSuEG3KTTGR')
         })
       })
     })