+++ CHANGELOG.md
@@ -2,7 +2,7 @@ next version / 2014-06-xx
 -------------------------
 * added `mochify` as a dev dep
 * removed semicolons per http://cryptocoinjs.com/about/contributing/#semicolons
-
+* removed `terst` and moved to `assert`
 
 0.1.0 / 2014-03-10
 ------------------